ov002: promote daObjMarioCap_c into a single C++ TU (29 of 30)
Mario's knocked-off cap becomes one production source,
src/actors/daObjMarioCap_c.cpp, absorbing 29 of the 30 legacy one-function
shards that used to carry it. The ROM build takes all 29 from a single object,
the way the cartridge's own build did.

WHAT THE CARTRIDGE PROVES ABOUT THE CLASS

  _ZTS  ov002 0x021095b8  "15daObjMarioCap_c"
  _ZTI  ov002 0x021095ac  __si_class_type_info; +8 -> _ZTI12dEnemyBase_c
                          (ov002 0x021081c0), so the DIRECT base is
                          dEnemyBase_c and nothing else
  _ZTV  ov002 0x021095f0  the address point; the emitted object starts two
                          words earlier at 0x021095e8, 0x84 bytes = 2 preamble
                          words + 31 slots, dActor_c's table length
  size  0x410             the factory's own literal (1040)

The vtable was diffed word for word against _ZTV12dEnemyBase_c (ov002
0x021081e4). Exactly nine slots differ -- 0 InitResources, 3 CleanupResources,
6 Behavior, 9 Render, 12 OnPendingDestroy, the 16/17 destructor pair, 18
OnYoshiTryEat and 19 OnTurnIntoEgg -- and they are exactly the nine methods
include/daObjMarioCap_c.h declares. Every other slot holds dEnemyBase_c's own
word, so nothing else is overridden and nothing else is redeclared.

WHY 29 OF 30, AND WHICH ONE IS HELD OUT

_ZN15daObjMarioCap_c13InitResourcesEv (0x020b86d0, size 0x4c8) does not
reproduce. src/_ZN15daObjMarioCap_c13InitResourcesEv.cpp compiles to 0x4d0
under the pinned 2004/b56 -- two instructions over -- and its delinks.txt block
never carried a `complete` marker. This is the promotion queue's recorded
`unmatched:1` blocker, measured rather than inherited.

It is the TOP function of the run, so holding it out costs no hole: the
licensed claim is the contiguous 0x020b6f18..0x020b86d0, 6072 bytes, and the
shard stays enrolled as its own delinks entry above it. A licensed claim cannot
span a gap, so a sourceless or unmatched function in the MIDDLE would have cost
far more; this one is free.

THE DESTRUCTOR IS INLINE AND DECLARED FIRST. The cartridge puts D1 at
0x020b6f18 BELOW D0 at 0x020b6f68 and carries no D2 anywhere in the image --
the reproducible direction. Inline in the class emits exactly that pair in
exactly that order; out of line emits D2/D0/D1 plus a homeless D2. The typed
member list makes the empty body own the dCapIcon_c, ShadowModel, ModelAnim,
dBgCh_Actr and dCcAc_c teardowns and the chain into _ZN12dEnemyBase_cD2Ev.

Making it inline moves the key function to OnYoshiTryEat -- the first
out-of-line virtual the class declares -- which this TU defines, so the vtable
and RTTI group are still emitted here.

TWELVE COMPILER-ONLY RECORDS, ZERO EXTERNALIZED. 2 x (4 ancestors + self) = 10
vague-linkage RTTI records, plus the vtable, plus _ZN7Vector3D1Ev: types.h
gives Vector3 an inline destructor and two members hold Vector3 locals, so the
TU odr-uses it and emits its own four-byte copy. All twelve addresses have a
configured ROM home in some module's symbols.txt (seven in arm9, five in
ov002), so all twelve license as deadstrip-data / deadstrip-duplicate and
externalized_output is empty.

RECONCILING THE MERGE. tubuild create reported 40 declaration conflicts and 3
RAW blocks over a preamble that did not compile (a declaration split across a
comment, two contradictory PMF typedefs, and data_ov002_0210de30 spelled three
mutually exclusive ways). The generated preamble was discarded and hand-curated
to one declaration per symbol, with casts at the call sites that needed them.
Two merge hazards were headed off rather than debugged: math/Matrix.h reaches
this TU through ModelAnim.h and spells Matrix4x3 as `Matrix3x3 r; Vector3 t`,
so func_ov002_020b7f7c's whole-matrix assignments use a local flat Mat12; and
func_ov002_020b7e1c's shadow fBase_c/dActor_c/SaveData structs were replaced
by the real headers plus mangled externs.

PROOF

  tools/tubuild.py verify ov002/daObjMarioCap_c
    29/29 MATCH, objisolate clean, reloc-destinations clean -> TEXT-VERIFIED
    0 unlicensed section/symbol(s)
  tools/tubuild.py linkcheck ov002/daObjMarioCap_c   (pre-promotion)
    SCRATCH-LINK-VERIFIED; TU .text range 0x020b6f18..0x020b86d0, 6072 bytes,
    IDENTICAL; module fidelity 106/106 exact, 100.000000%; ROM sha256
    d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8,
    identical to the baseline control
    [7/8] dsd check symbols --fail: 9 pre-existing errors, 0 new (the baseline
    control has the same 9)
  static gates: layout_check, check_dead_references, check_duplicate_sources,
    check_tubuild_conflicts, check_src_tu_compiles, port_refcheck,
    check_header_offsets, cpp_tu_state --check-note all clean
  tools/tiers_ratchet.py --update: 2680 -> 2680 (+3 / -3), the three removals
    banked as the consolidation they are

Full-ROM proof is the builder's stage and was deliberately not run here.

BUILDER (rebased three times as main moved: 66779c216 -> e7c6f7ce1 ->
395afdb23 -> d8b33da4d, 25 commits in all). The
compiler-only split is seven records in arm9 and five in ov002, not eight and
four; notes/data/class-facts/daObjMarioCap_c.json carried the same slip and is
corrected here.

Only two files ever conflicted across the three rebases, and neither is source:
notes/cpp-tu-current-state.md, which is generated and was regenerated rather
than resolved, and attribution.json, resolved line-wise. Both ledger files
auto-merged every time, and the merge was audited correct by path-multiset diff
against origin/main -- exceptions 346 -> 349 (+3 ours, 0 lost, 0 byte-identical
duplicates), CONVERTED set 2681 -> 2681 with the three shard paths replaced by
three path#symbol entries. tiers_ratchet was therefore NOT re-run: --check
passes at 2681/2681 on the new base.

notes/data/tu-promotion-queue.tsv: this class's row is refreshed to
already_promoted=yes, shard_count 31 -> 3, total_lines 1278 -> 1295, matching
what tools/queue_audit.py (landed in #2320) derives. The row was edited in
place rather than taking that tool's --write output wholesale: --write
round-trips the file and silently drops four header comment lines (263 lines
out against 267 in), and it would also have rewritten dScMgBomroom_c's row,
which belongs to #2319. The hand-edited row is byte-identical to the one
--write derives. queue_audit is not wired into any workflow.

attribution.json carries 26 src/actors/daObjMarioCap_c.cpp#<symbol> override
rows so the fold keeps each member's matcher. Resolved line-wise on every
rebase, never re-serialised; the union against main is exact (main's rows + 26,
0 lost, 0 changed) and none of the 26 credits github-actions[bot].
