+++ attribution.json
@@ -2347,6 +2347,32 @@
     "src/actors/Goomboss.cpp#func_ov074_02120d74": "tangosdev",
     "src/actors/Goomboss.cpp#func_ov074_021201f0": "tangosdev",
     "src/actors/Goomboss.cpp#func_ov074_021201ec": "tangosdev",
-    "src/actors/Goomboss.cpp#func_ov074_0211ffac": "tangosdev"
+    "src/actors/Goomboss.cpp#func_ov074_0211ffac": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7f24": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b6fcc": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7c30": "ruspecial",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b71f0": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7f2c": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7cec": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7f7c": "ruspecial",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7d58": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b74d0": "noreply",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7e1c": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b76ec": "andrewboudreau",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7cdc": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_cD1Ev": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_c8BehaviorEv": "noreply",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_c6RenderEv": "andrewboudreau",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_c13OnTurnIntoEggER6Player": "andrewboudreau",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7e08": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7d9c": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b71e8": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_c16CleanupResourcesEv": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b781c": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7d6c": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_c16OnPendingDestroyEv": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7330": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#func_ov002_020b7d94": "tangosdev",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_c13OnYoshiTryEatEv": "andrewboudreau"
   }
 }
