+++ config/arm9/overlays/ov002/delinks.txt
@@ -845,121 +845,9 @@ src/game/actors/d_a_obj_waterfall.cpp:
     complete
     .text start:0x020b6e08 end:0x020b6f18
 
-src/_ZN15daObjMarioCap_cD1Ev.cpp:
+src/actors/daObjMarioCap_c.cpp:
     complete
-    .text start:0x020b6f18 end:0x020b6f68
-
-src/_ZN15daObjMarioCap_cD0Ev.cpp:
-    complete
-    .text start:0x020b6f68 end:0x020b6fcc
-
-src/func_ov002_020b6fcc.cpp:
-    complete
-    .text start:0x020b6fcc end:0x020b71e8
-
-src/func_ov002_020b71e8.c:
-    complete
-    .text start:0x020b71e8 end:0x020b71f0
-
-src/func_ov002_020b71f0.c:
-    complete
-    .text start:0x020b71f0 end:0x020b7200
-
-src/func_ov002_020b7200.c:
-    complete
-    .text start:0x020b7200 end:0x020b7330
-
-src/func_ov002_020b7330.c:
-    complete
-    .text start:0x020b7330 end:0x020b74d0
-
-src/func_ov002_020b74d0.c:
-    complete
-    .text start:0x020b74d0 end:0x020b76ec
-
-src/func_ov002_020b76ec.c:
-    complete
-    .text start:0x020b76ec end:0x020b781c
-
-src/func_ov002_020b781c.c:
-    complete
-    .text start:0x020b781c end:0x020b7b70
-
-src/func_ov002_020b7b70.c:
-    complete
-    .text start:0x020b7b70 end:0x020b7c30
-
-src/func_ov002_020b7c30.c:
-    complete
-    .text start:0x020b7c30 end:0x020b7cdc
-
-src/func_ov002_020b7cdc.c:
-    complete
-    .text start:0x020b7cdc end:0x020b7cec
-
-src/func_ov002_020b7cec.cpp:
-    complete
-    .text start:0x020b7cec end:0x020b7d58
-
-src/func_ov002_020b7d58.c:
-    complete
-    .text start:0x020b7d58 end:0x020b7d6c
-
-src/func_ov002_020b7d6c.c:
-    complete
-    .text start:0x020b7d6c end:0x020b7d94
-
-src/func_ov002_020b7d94.c:
-    complete
-    .text start:0x020b7d94 end:0x020b7d9c
-
-src/func_ov002_020b7d9c.cpp:
-    complete
-    .text start:0x020b7d9c end:0x020b7e08
-
-src/func_ov002_020b7e08.c:
-    complete
-    .text start:0x020b7e08 end:0x020b7e1c
-
-src/func_ov002_020b7e1c.cpp:
-    complete
-    .text start:0x020b7e1c end:0x020b7f24
-
-src/func_ov002_020b7f24.c:
-    complete
-    .text start:0x020b7f24 end:0x020b7f2c
-
-src/func_ov002_020b7f2c.cpp:
-    complete
-    .text start:0x020b7f2c end:0x020b7f7c
-
-src/func_ov002_020b7f7c.c:
-    complete
-    .text start:0x020b7f7c end:0x020b81e0
-
-src/_ZN15daObjMarioCap_c13OnTurnIntoEggER6Player.cpp:
-    complete
-    .text start:0x020b81e0 end:0x020b8270
-
-src/_ZN15daObjMarioCap_c13OnYoshiTryEatEv.cpp:
-    complete
-    .text start:0x020b8270 end:0x020b8284
-
-src/_ZN15daObjMarioCap_c16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x020b8284 end:0x020b8394
-
-src/_ZN15daObjMarioCap_c16OnPendingDestroyEv.cpp:
-    complete
-    .text start:0x020b8394 end:0x020b83c4
-
-src/_ZN15daObjMarioCap_c6RenderEv.cpp:
-    complete
-    .text start:0x020b83c4 end:0x020b84a8
-
-src/_ZN15daObjMarioCap_c8BehaviorEv.cpp:
-    complete
-    .text start:0x020b84a8 end:0x020b86d0
+    .text start:0x020b6f18 end:0x020b86d0
 
 src/_ZN15daObjMarioCap_c13InitResourcesEv.cpp:
     .text start:0x020b86d0 end:0x020b8b98
