+++ config/converted-backslide-exceptions.jsonl
@@ -348,3 +348,6 @@
 {"path": "src/_ZN13dScMgTeresa_cD1Ev.cpp", "reason": "ov006/dScMgTeresa_c promotion: 48 one-function shards absorbed into src/actors/dScMgTeresa_c.cpp (ROM ordinals 0..47 of the 0x0211cbd0..0x021207a8 run; the other 33 stay enrolled behind a sourceless hole)"}
 {"path": "src/_ZN6Eyerok16CleanupResourcesEv.cpp", "reason": "ov066/Eyerok TU promotion: 34 of the 61-function run at 0x02115ee0..0x0211a2e4 fold into src/actors/Eyerok.cpp (.text 0x021184c0..0x02119ce8). Partial because six shards in the run reference cross-overlay addresses no symbols.txt names and so carry no complete marker; a licensed claim cannot contain a hole."}
 {"path": "src/_ZN6Eyerok16OnPendingDestroyEv.cpp", "reason": "ov066/Eyerok TU promotion: 34 of the 61-function run at 0x02115ee0..0x0211a2e4 fold into src/actors/Eyerok.cpp (.text 0x021184c0..0x02119ce8). Partial because six shards in the run reference cross-overlay addresses no symbols.txt names and so carry no complete marker; a licensed claim cannot contain a hole."}
+{"path": "src/_ZN15daObjMarioCap_c16CleanupResourcesEv.cpp", "reason": "ov002/daObjMarioCap_c TU promotion: 29 one-function shards absorbed into src/actors/daObjMarioCap_c.cpp, 29/29 byte-verified. The three readability criteria that regress are required by the byte match, not chosen: the 21 func_ov002_020b* members occupy no vtable slot and the ROM names none of them, so no real name exists to give them; their bodies reach the object through raw offsets because that is the spelling the cartridge's bytes reproduce (naming a field lets mwcc CSE the field address and the function comes out short); and a reconstructed TU must spell vague-linkage and by-value-Fix12 symbols by their mangled names (_ZN8dActor_c9SetRangesE5Fix12IiES1_S1_S1_, _ZN9ModelAnim7SetAnimEP8BCA_Filei5Fix12IiEj) or the range will not link at all."}
+{"path": "src/_ZN15daObjMarioCap_cD0Ev.cpp", "reason": "ov002/daObjMarioCap_c TU promotion: 29 one-function shards absorbed into src/actors/daObjMarioCap_c.cpp, 29/29 byte-verified. The three readability criteria that regress are required by the byte match, not chosen: the 21 func_ov002_020b* members occupy no vtable slot and the ROM names none of them, so no real name exists to give them; their bodies reach the object through raw offsets because that is the spelling the cartridge's bytes reproduce (naming a field lets mwcc CSE the field address and the function comes out short); and a reconstructed TU must spell vague-linkage and by-value-Fix12 symbols by their mangled names (_ZN8dActor_c9SetRangesE5Fix12IiES1_S1_S1_, _ZN9ModelAnim7SetAnimEP8BCA_Filei5Fix12IiEj) or the range will not link at all."}
+{"path": "src/_ZN15daObjMarioCap_cD1Ev.cpp", "reason": "ov002/daObjMarioCap_c TU promotion: 29 one-function shards absorbed into src/actors/daObjMarioCap_c.cpp, 29/29 byte-verified. The three readability criteria that regress are required by the byte match, not chosen: the 21 func_ov002_020b* members occupy no vtable slot and the ROM names none of them, so no real name exists to give them; their bodies reach the object through raw offsets because that is the spelling the cartridge's bytes reproduce (naming a field lets mwcc CSE the field address and the function comes out short); and a reconstructed TU must spell vague-linkage and by-value-Fix12 symbols by their mangled names (_ZN8dActor_c9SetRangesE5Fix12IiES1_S1_S1_, _ZN9ModelAnim7SetAnimEP8BCA_Filei5Fix12IiEj) or the range will not link at all."}
