+++ config/converted-baseline.json
@@ -838,9 +838,6 @@
     "src/_ZN15daObjGuragura_c6RenderEv.cpp",
     "src/_ZN15daObjGuragura_cD0Ev.cpp",
     "src/_ZN15daObjGuragura_cD1Ev.cpp",
-    "src/_ZN15daObjMarioCap_c16CleanupResourcesEv.cpp",
-    "src/_ZN15daObjMarioCap_cD0Ev.cpp",
-    "src/_ZN15daObjMarioCap_cD1Ev.cpp",
     "src/_ZN16BowserShockwaves16CleanupResourcesEv.cpp",
     "src/_ZN16BowserShockwaves6RenderEv.cpp",
     "src/_ZN16BowserShockwavesD0Ev.cpp",
@@ -2165,6 +2162,9 @@
     "src/actors/daObjFallBlock_c.cpp#_ZN16daObjFallBlock_c6RenderEv",
     "src/actors/daObjFloatBoard_c.cpp#_ZN17daObjFloatBoard_c16CleanupResourcesEv",
     "src/actors/daObjFloatBoard_c.cpp#_ZN17daObjFloatBoard_c6RenderEv",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_c16CleanupResourcesEv",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_cD0Ev",
+    "src/actors/daObjMarioCap_c.cpp#_ZN15daObjMarioCap_cD1Ev",
     "src/actors/daOts_c.cpp#_ZN7daOts_c12PlayHitSoundEv",
     "src/actors/daOts_c.cpp#_ZN7daOts_c13PlayStepSoundEv",
     "src/actors/daOts_c.cpp#_ZN7daOts_c16OnAimedAtWithEggEv",
