+++ config/match_attempts.jsonl
@@ -1921,7 +1921,7 @@
 {"schemaVersion":1,"functionId":"ov063:0x02115fc4","id":"ov063:0x02115fc4","attemptId":"9c8ce254d6ff44b3a3b6802f1dd1b8b2","parentAttemptId":null,"module":"ov063","addr":34693060,"name":"_ZN11daTBasket_cD1Ev","status":"matched","kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex","author":"tangosdev","improvedNearMiss":false,"srcPath":"src/actors/BooCage/_ZN11daTBasket_cD1Ev.cpp","label":"datbasket-native-destructors","note":"ROM RTTI corrects the imported BooCage lifecycle identity to daTBasket_c; the native D1 generates ShadowModel, Model, dBgCh_Actr, dCcAc_c, and dEnemyBase_c teardown in ROM order.","base":{"kind":"matched_sibling"},"sessionScope":"batch","batchSize":2}
 {"schemaVersion":1,"functionId":"ov063:0x0211600c","id":"ov063:0x0211600c","attemptId":"c7991adfb8b84ed7b4ce99e83e5e64db","parentAttemptId":null,"module":"ov063","addr":34693132,"name":"_ZN11daTBasket_cD0Ev","status":"matched","kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex","author":"tangosdev","improvedNearMiss":false,"srcPath":"src/actors/BooCage/_ZN11daTBasket_cD0Ev.cpp","label":"datbasket-native-destructors","note":"The real dEnemyBase_c inheritance surface makes CodeWarrior append the actor-heap deallocation sequence to the generated member/base teardown; VERIFIED with blind 0.","base":{"kind":"matched_sibling"},"sessionScope":"batch","batchSize":2}
 {"schemaVersion":1,"functionId":"ov063:0x0211d3f0","id":"ov063:0x0211d3f0","attemptId":"d49c7bd3b52d48fca40b6b6599350d03","parentAttemptId":null,"module":"ov063","addr":34722800,"name":"_ZN12FallBlockBbhD0Ev","status":"matched","kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex","author":"tangosdev","improvedNearMiss":false,"srcPath":"src/actors/FallBlockBbh/_ZN12FallBlockBbhD0Ev.cpp","label":"fallblockbbh-native-deleting-destructor","note":"The same empty native destructor already proven by D1 makes CodeWarrior generate all three vptr transitions, inherited member/base teardown, and actor-heap deallocation for D0; VERIFIED with blind 0.","base":{"kind":"matched_sibling"},"sessionScope":"single","batchSize":1}
-{"schemaVersion":1,"functionId":"ov002:0x020b6f68","id":"ov002:0x020b6f68","attemptId":"dae2da96c6904ad7a160849ca7b3f2e6","parentAttemptId":null,"module":"ov002","addr":34303848,"name":"_ZN15daObjMarioCap_cD0Ev","status":"matched","kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex","author":"tangosdev","improvedNearMiss":false,"srcPath":"src/_ZN15daObjMarioCap_cD0Ev.cpp","label":"mariocap-native-deleting-destructor","note":"The same empty native destructor already proven by D1 makes CodeWarrior generate five typed member teardowns, dEnemyBase_c teardown, and actor-heap deallocation for D0; VERIFIED with blind 0.","base":{"kind":"matched_sibling"},"sessionScope":"single","batchSize":1}
+{"schemaVersion": 1, "functionId": "ov002:0x020b6f68", "id": "ov002:0x020b6f68", "attemptId": "dae2da96c6904ad7a160849ca7b3f2e6", "parentAttemptId": null, "module": "ov002", "addr": 34303848, "name": "_ZN15daObjMarioCap_cD0Ev", "status": "matched", "kind": "ai", "model": "gpt-5", "reasoning": "high", "harness": "codex", "author": "tangosdev", "improvedNearMiss": false, "srcPath": "src/actors/daObjMarioCap_c.cpp", "label": "mariocap-native-deleting-destructor", "note": "The same empty native destructor already proven by D1 makes CodeWarrior generate five typed member teardowns, dEnemyBase_c teardown, and actor-heap deallocation for D0; VERIFIED with blind 0.", "base": {"kind": "matched_sibling"}, "sessionScope": "single", "batchSize": 1}
 {"schemaVersion":1,"functionId":"ov002:0x020aedbc","id":"ov002:0x020aedbc","attemptId":"cc85909c82134de191b37acf9410890c","parentAttemptId":null,"module":"ov002","addr":34270652,"name":"_ZN11dCapEnemy_cD1Ev","status":"matched","kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex","author":"tangosdev","improvedNearMiss":false,"srcPath":"src/_ZN11dCapEnemy_cD1Ev.cpp","label":"capenemy-native-overlay-destructors","note":"Vtable slot 16 and two derived base chains identify the anonymous overlay function as dCapEnemy_c D1; the native destructor already proven by ARM9 D2 reproduces it with blind 0.","base":{"kind":"matched_sibling"},"sessionScope":"batch","batchSize":2}
 {"schemaVersion":1,"functionId":"ov002:0x020aedf4","id":"ov002:0x020aedf4","attemptId":"f8c86fe0559e42b59de8c71a6d7bcd73","parentAttemptId":null,"module":"ov002","addr":34270708,"name":"_ZN11dCapEnemy_cD0Ev","status":"matched","kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex","author":"tangosdev","improvedNearMiss":false,"srcPath":"src/_ZN11dCapEnemy_cD0Ev.cpp","label":"capenemy-native-overlay-destructors","note":"The same native destructor generates CapIcon, Model, dEnemyBase_c, and inherited actor-heap teardown for overlay D0; VERIFIED with blind 0.","base":{"kind":"matched_sibling"},"sessionScope":"batch","batchSize":2}
 {"schemaVersion":1,"functionId":"ov031:0x021111d0","id":"ov031:0x021111d0","attemptId":"b4d3d9e6f91f4c2bb16ab6df8ff72f93","parentAttemptId":null,"module":"ov031","addr":34673104,"name":"_ZN25SlideDecorationSilverStarD0Ev","status":"matched","kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex","author":"tangosdev","improvedNearMiss":false,"srcPath":"src/_ZN25SlideDecorationSilverStarD0Ev.cpp","label":"slide-decoration-native-destructors","note":"The real dActor_c-derived class, explicit actor tail gap, owned Model, and inherited operator delete make CodeWarrior generate the complete D0 body; VERIFIED with blind 0.","base":{"kind":"matched_sibling"},"sessionScope":"single","batchSize":1}
