+++ config/match_provenance.jsonl
@@ -497,7 +497,7 @@
 {"id":"ov063:0x0211c444","module":"ov063","addr":34718788,"name":"_ZN11daTrsIcon_c13InitResourcesEv","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/actors/BigBooIcon/_ZN11daTrsIcon_c13InitResourcesEv.cpp"}
 {"id":"ov064:0x0211a968","module":"ov064","addr":34711912,"name":"_ZN4ClamD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN4ClamD0Ev.cpp"}
 {"id":"ov063:0x0211d3f0","module":"ov063","addr":34722800,"name":"_ZN12FallBlockBbhD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/actors/FallBlockBbh/_ZN12FallBlockBbhD0Ev.cpp"}
-{"id":"ov002:0x020b6f68","module":"ov002","addr":34303848,"name":"_ZN15daObjMarioCap_cD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN15daObjMarioCap_cD0Ev.cpp"}
+{"id": "ov002:0x020b6f68", "module": "ov002", "addr": 34303848, "name": "_ZN15daObjMarioCap_cD0Ev", "matchProvenance": {"kind": "ai", "model": "gpt-5", "reasoning": "high", "harness": "codex"}, "author": "tangosdev", "srcPath": "src/actors/daObjMarioCap_c.cpp"}
 {"id":"ov002:0x020aedbc","module":"ov002","addr":34270652,"name":"_ZN11dCapEnemy_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN11dCapEnemy_cD1Ev.cpp"}
 {"id":"ov002:0x020aedf4","module":"ov002","addr":34270708,"name":"_ZN11dCapEnemy_cD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN11dCapEnemy_cD0Ev.cpp"}
 {"id":"ov031:0x021111d0","module":"ov031","addr":34673104,"name":"_ZN25SlideDecorationSilverStarD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN25SlideDecorationSilverStarD0Ev.cpp"}
@@ -553,7 +553,7 @@
 {"id":"ov001:0x020ab374","module":"ov001","addr":34255732,"name":"_ZN10dCapIcon_cD0Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN10dCapIcon_cD0Ev.cpp"}
 {"id":"ov001:0x020ab3a0","module":"ov001","addr":34255776,"name":"_ZN10dCapIcon_cD1Ev","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN10dCapIcon_cD1Ev.cpp"}
 {"id":"ov065:0x021190a8","module":"ov065","addr":34705576,"name":"_ZN12daDossyCap_c8BehaviorEv","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN12daDossyCap_c8BehaviorEv.cpp"}
-{"id":"ov002:0x020b8394","module":"ov002","addr":34309012,"name":"_ZN15daObjMarioCap_c16OnPendingDestroyEv","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/_ZN15daObjMarioCap_c16OnPendingDestroyEv.cpp"}
+{"id": "ov002:0x020b8394", "module": "ov002", "addr": 34309012, "name": "_ZN15daObjMarioCap_c16OnPendingDestroyEv", "matchProvenance": {"kind": "ai", "model": "gpt-5", "reasoning": "high", "harness": "codex"}, "author": "tangosdev", "srcPath": "src/actors/daObjMarioCap_c.cpp"}
 {"id":"arm9:0x02005ed8","module":"arm9","addr":33578712,"name":"func_02005ed8","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/func_02005ed8.c"}
 {"id":"ov001:0x020aa420","module":"ov001","addr":34251808,"name":"func_ov001_020aa420","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/func_ov001_020aa420.c"}
 {"id":"ov001:0x020aaa54","module":"ov001","addr":34253396,"name":"func_ov001_020aaa54","matchProvenance":{"kind":"ai","model":"gpt-5","reasoning":"high","harness":"codex"},"author":"tangosdev","srcPath":"src/func_ov001_020aaa54.cpp"}
