+++ notes/actor-core-provenance.md
@@ -136,9 +136,13 @@ base. A declaration on `fBase_c` does NOT reach `dActor_c` (two levels down), no
 classes `fBase_c`'s copy changes are `fBase_c` itself and `dBase_c` -- the two
 whose D0 the ROM shows inlining it. `dEnemyBase_c` carries its own copy for the
 same reason: it is a flattened struct that does not derive from `dActor_c` in
-these headers. The two `src/` files that declare a local `struct dActor_c : fBase_c`
-(`EndKuppaScript.cpp`, [func_ov002_020b7e1c.cpp](../src/func_ov002_020b7e1c.cpp)) use their own local shadow
-`fBase_c`, not this one, so they are out of scope too.
+these headers. The one `src/` file that still declares a local
+`struct dActor_c : fBase_c` (`EndKuppaScript.cpp`) uses its own local shadow
+`fBase_c`, not this one, so it is out of scope too. There used to be a second,
+the one-function shard behind `func_ov002_020b7e1c`; the daObjMarioCap_c
+promotion folded it into
+[src/actors/daObjMarioCap_c.cpp](../src/actors/daObjMarioCap_c.cpp), where the
+real `dActor_c` and `fBase_c` are in scope and the shadow is gone.
 
 No layout effect: an inline non-virtual member adds no field and no vtable slot,
 and the `0x50` / `0xd0` size assertions still hold.
