--- src/_ZN15daObjMarioCap_c13OnYoshiTryEatEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN15daObjMarioCap_c13OnYoshiTryEatEv
-#include "daObjMarioCap_c.h"
-/* recovered: renamed to Class_Method */
-s32 daObjMarioCap_c::OnYoshiTryEat() {
-    char* c = (char*)this;
-  if(*(int*)(c+0x3f0)==2) return 0;
-  return 4;
-}
