--- src/_ZN15daObjMarioCap_cD0Ev.cpp
@@ -1,12 +1,0 @@
-//cpp
-// @symbol _ZN15daObjMarioCap_cD0Ev
-/* recovered: native deleting destructor -- the compiler emits the whole body
- *
- * The typed class layout makes the empty body own the dCapIcon_c, ShadowModel,
- * ModelAnim, dBgCh_Actr, dCcAc_c, dEnemyBase_c, and actor-heap teardown.
- */
-#include "daObjMarioCap_c.h"
-
-daObjMarioCap_c::~daObjMarioCap_c()
-{
-}
