--- src/_ZN15daObjMarioCap_cD1Ev.cpp
@@ -1,13 +1,0 @@
-//cpp
-// @symbol _ZN15daObjMarioCap_cD1Ev
-/* recovered: real C++ destructor -- the compiler emits the whole body
- *
- * One vtable store and five member teardowns in reverse declaration order,
- * followed by dEnemyBase_c. The first member teardown is the ROM-proven
- * dCapIcon_c destructor at 0x020ab3a0.
- */
-#include "daObjMarioCap_c.h"
-
-daObjMarioCap_c::~daObjMarioCap_c()
-{
-}
