--- src/func_ov002_020b71e8.c
@@ -1,4 +1,0 @@
-int func_ov002_020b71e8(void)
-{
-    return 1;
-}
