--- src/func_ov002_020b71f0.c
@@ -1,4 +1,0 @@
-int func_ov002_020b71f0(int *p)
-{
-    p[251] = 0; return 1;
-}
