--- src/func_ov002_020b7cdc.c
@@ -1,8 +1,0 @@
-// @symbol func_ov002_020b7cdc
-// recovered name: PoppingLavaBubbles_Kill
-/* recovered: renamed to Class_Method */
-/* daObjWaterfall_c::Kill - recovered from vtable slot identity */
-int func_ov002_020b7cdc(int *p)
-{
-    p[39] = 0; return 1;
-}
