--- src/func_ov002_020b7d94.c
@@ -1,4 +1,0 @@
-int func_ov002_020b7d94(void)
-{
-    return 1;
-}
