--- src/func_ov002_020b7e08.c
@@ -1,5 +1,0 @@
-int func_ov002_020b7e08(char *p)
-{
-    *(short *)(p + 0x100) = 200;
-    return 1;
-}
