--- src/func_ov002_020b7f24.c
@@ -1,4 +1,0 @@
-int func_ov002_020b7f24(void)
-{
-    return 1;
-}
